Big update. Mutators are functioning very well here now. Also minor bug fixes to other areas of the Peach internals.

+++ Peach.Core.Proxy/Web/HttpMessage.cs
@@ -88,7 +88,7 @@ namespace Peach.Core.Proxy.Web
 
 		protected static string ReadLine(BitStream data)
 		{
-			int currentPosition = data.TellBytes();
+			int currentPosition = (int)data.TellBytes();
 
 			try
 			{
